Move module helpers out of Eval
