+++ lib/eval.js
@@ -82,51 +82,6 @@ class Eval {
     }, filePath)
   }
 
-  static getRenderRequireGraph () {
-    return Eval.execute(() => {
-      let count = 0
-      const walkModule = (module) => {
-        count++
-        let modulePath = module.filename || module.id
-        if (process.platform === 'win32') {
-          modulePath = modulePath.replace(/\\/g, '/')
-        }
-        return {
-          path: modulePath,
-          children: module.children.map(walkModule)
-        }
-      }
-      const mainModule = walkModule(process.mainModule)
-      mainModule.resourcesPath = process.resourcesPath
-      mainModule.appName = require('electron').remote.app.getName()
-      mainModule.count = count
-      return mainModule
-    })
-  }
-
-  static getMainRequireGraph () {
-    return Eval.execute(() => {
-      let process = require('electron').remote.process
-      let count = 0
-      const walkModule = (module) => {
-        count++
-        let modulePath = module.filename || module.id
-        if (process.platform === 'win32') {
-          modulePath = modulePath.replace(/\\/g, '/')
-        }
-        return {
-          path: modulePath,
-          children: module.children.map(walkModule)
-        }
-      }
-      const mainModule = walkModule(process.mainModule)
-      mainModule.resourcesPath = process.resourcesPath
-      mainModule.appName = require('electron').remote.app.getName()
-      mainModule.count = count
-      return mainModule
-    })
-  }
-
   static isDebugMode () {
     return Eval.execute(() => {
       return !!process.env.DEVTRON_DEBUG_PATH
