+++ lib/module-helpers.js
@@ -38,10 +38,55 @@ const createModules = (mainModule) => {
   return convertedMainModule
 }
 
+const getRenderRequireGraph = () => {
+  return Eval.execute(() => {
+    let count = 0
+    const walkModule = (module) => {
+      count++
+      let modulePath = module.filename || module.id
+      if (process.platform === 'win32') {
+        modulePath = modulePath.replace(/\\/g, '/')
+      }
+      return {
+        path: modulePath,
+        children: module.children.map(walkModule)
+      }
+    }
+    const mainModule = walkModule(process.mainModule)
+    mainModule.resourcesPath = process.resourcesPath
+    mainModule.appName = require('electron').remote.app.getName()
+    mainModule.count = count
+    return mainModule
+  })
+}
+
+const getMainRequireGraph = () => {
+  return Eval.execute(() => {
+    let process = require('electron').remote.process
+    let count = 0
+    const walkModule = (module) => {
+      count++
+      let modulePath = module.filename || module.id
+      if (process.platform === 'win32') {
+        modulePath = modulePath.replace(/\\/g, '/')
+      }
+      return {
+        path: modulePath,
+        children: module.children.map(walkModule)
+      }
+    }
+    const mainModule = walkModule(process.mainModule)
+    mainModule.resourcesPath = process.resourcesPath
+    mainModule.appName = require('electron').remote.app.getName()
+    mainModule.count = count
+    return mainModule
+  })
+}
+
 exports.getRenderModules = () => {
-  return Eval.getRenderRequireGraph().then(createModules).then(loadSizes).then(loadVersions)
+  return getRenderRequireGraph().then(createModules).then(loadSizes).then(loadVersions)
 }
 
 exports.getMainModules = () => {
-  return Eval.getMainRequireGraph().then(createModules).then(loadSizes).then(loadVersions)
+  return getMainRequireGraph().then(createModules).then(loadSizes).then(loadVersions)
 }
